Added CX mirror pos vectors
Added the vector function describing convex mirror form, to rotate the mirror at center of curve point multiply every column vector stored in the form matrix of the mirror with the rotation matrix. Translations can be done similarly by adding a repeated translation vector (repeated into a matrix) and adding it to the form matrix. The reflections on the mirror can be calculated using the original vector function and finding the intersection point with the incoming ray. (Maybe no need for form matrix if I can do the reflection calculations just with the vector function itself?)